+++ src/i18n/locales/en.js
@@ -1,6 +1,8 @@
 // This is just an example,
 // so you can safely delete all default props below
 
+import { info } from "tauri-plugin-log-api";
+
 export default {
     general: {
         failed: 'Action failed',
@@ -52,6 +54,11 @@ export default {
         },
         delete: {
             tooltip:'Delete'
+        },
+        info: {
+            tooltip:'Info',
+            title: 'Conversation info',
+            usage: 'Total token usage'
         },
     },